Revert "fix(control-workspace): correct Whisper link to open More Feature instead of mWeb"
