+++ babel.config.js
@@ -46,7 +46,6 @@ const defaultPluginsForWebpack = [
 
     // Keep it last
     'react-native-reanimated/plugin',
-    '@babel/plugin-transform-export-namespace-from',
 ];
 
 // The Fullstory annotate plugin generated a few errors when executed in Electron. Let's
@@ -133,7 +132,6 @@ const metro = {
                 },
             },
         ],
-        '@babel/plugin-transform-export-namespace-from',
     ],
     env: {
         production: {
