+++ package.json
@@ -130,7 +130,7 @@
     "date-fns": "^4.1.0",
     "date-fns-tz": "^3.2.0",
     "dom-serializer": "^0.2.2",
-    "domhandler": "^5.0.3",
+    "domhandler": "^4.3.0",
     "expensify-common": "2.0.156",
     "expo": "53.0.7",
     "expo-asset": "^11.1.2",
@@ -144,7 +144,7 @@
     "group-ib-fp": "file:modules/group-ib-fp",
     "heic-to": "^1.1.13",
     "howler": "^2.2.4",
-    "htmlparser2": "10.0.0",
+    "htmlparser2": "^7.2.0",
     "idb-keyval": "^6.2.1",
     "jszip": "^3.10.1",
     "lodash-es": "4.17.21",
@@ -221,7 +221,6 @@
     "@babel/plugin-proposal-private-methods": "^7.18.6",
     "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
     "@babel/plugin-transform-class-properties": "^7.25.4",
-    "@babel/plugin-transform-export-namespace-from": "^7.27.1",
     "@babel/preset-env": "^7.25.3",
     "@babel/preset-flow": "^7.12.13",
     "@babel/preset-react": "^7.10.4",
