+++ src/hooks/useEnvironment.ts
@@ -1,6 +1,6 @@
 import {useContext} from 'react';
-import {EnvironmentContext} from '@components/EnvironmentContext';
-import type {EnvironmentContextValue} from '@components/EnvironmentContext';
+import {EnvironmentContext} from '@components/withEnvironment';
+import type {EnvironmentContextValue} from '@components/withEnvironment';
 import CONST from '@src/CONST';
 
 type UseEnvironment = EnvironmentContextValue & {
@@ -9,12 +9,11 @@ type UseEnvironment = EnvironmentContextValue & {
 };
 
 export default function useEnvironment(): UseEnvironment {
-    const {environment, environmentURL, adjustExpensifyLinksForEnv} = useContext(EnvironmentContext);
+    const {environment, environmentURL} = useContext(EnvironmentContext);
     return {
         environment,
         environmentURL,
         isProduction: environment === CONST.ENVIRONMENT.PRODUCTION,
         isDevelopment: environment === CONST.ENVIRONMENT.DEV,
-        adjustExpensifyLinksForEnv,
     };
 }
